requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic[email]
fastapi==0.95.0
aiogram==2.25.1
python-dotenv==1.0.0